requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-jose>=3.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
passlib>=1.7.4
bcrypt>=4.0.1
alembic>=1.13.0
python-multipart>=0.0.6
pytest-sugar>=0.9.7
